+++ src/main/java/com/busticket/respository/IAddressRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.busticket.entity.Address;
 
-public interface IAddressRepo extends JpaRepository<Address, Long>{
+public interface IAddressRepo extends JpaRepository<Address, Integer>{
 
 }
