+++ src/main/java/com/busticket/respository/IReviewRepo.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface IReviewRepo extends JpaRepository<Review, Long>{
+public interface IReviewRepo extends JpaRepository<Review, Integer>{
 
-   List<Review> findByCustomer_CustomerId(long customerId);
-   List<Review> findByTrip_TripId(long tripId);
+   List<Review> findByCustomer_CustomerId(Integer customerId);
+   List<Review> findByTrip_TripId(Integer tripId);
 }
