+++ src/main/java/com/busticket/respository/ITripRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.busticket.entity.Trip;
 
-public interface ITripRepo extends JpaRepository<Trip, Long>{
+public interface ITripRepo extends JpaRepository<Trip, Integer>{
 
 }
